requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.13.0
blinker==1.9.0
boto3==1.38.41
botocore==1.38.41
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.17.0
Flask==3.1.1
frozenlist==1.5.0
google==3.0.0
google-api-core==2.24.2
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-google-community==2.0.7
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langsmith==0.2.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
oauthlib==3.2.2
openai==1.59.9
orjson==3.10.15
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.13.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9
s3transfer==0.13.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.46.2
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.29.1
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
yarl==1.18.3
zipp==3.21.0